Turn essential template logic to hook and its own file
